+++ include/backend/impls/rng_cpu.h
