+++ src/backend/impls/rng_cpu.c
