+++ src/core/glwe/glwe_ciphertext.c
@@ -6,6 +6,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "backend_arithmetic.h"
 #include "bivariate_polynomial.h"
 #include "glwe_params.h"
 #include "maths_structures.h"
