+++ tests/unit/core/ggsw/test_trace_expand.c
@@ -8,7 +8,6 @@
 #include <sys/types.h>
 
 #include "bivariate_polynomial.h"
-#include "core/glwe/glwe_arithmetic.h"
 #include "core/glwe/glwe_ciphertext.h"
 #include "core/glwe/glwe_transform_key.h"
 #include "ggsw_arithmetic.h"
